refactor(log): extract rendering and summary parsing (#530)
