+++ lua/forge/log.lua
@@ -1,6 +1,8 @@
 local M = {}
 
 local config_mod = require('forge.config')
+local log_render = require('forge.log_render')
+local log_summary = require('forge.log_summary')
 local scope_mod = require('forge.scope')
 local ns = vim.api.nvim_create_namespace('forge_log')
 local buf_data = {}
@@ -93,598 +95,12 @@ local function request_current(buf, request_id)
   return d and d.request_id == request_id
 end
 
-local sgr_map = {
-  [1] = 'ForgeLogJob',
-  [2] = 'ForgeLogDim',
-  [31] = 'ForgeFail',
-  [32] = 'ForgePass',
-  [33] = 'ForgeLogWarning',
-  [34] = 'ForgeLogStep',
-  [36] = 'ForgeLogSection',
-  [242] = 'ForgeLogDim',
-  [91] = 'ForgeFail',
-  [92] = 'ForgePass',
-  [93] = 'ForgeLogWarning',
-  [96] = 'ForgeLogSection',
-}
-
-local function strip_ansi(line)
-  if line:byte(1) == 0xEF and line:byte(2) == 0xBB and line:byte(3) == 0xBF then
-    line = line:sub(4)
-  end
-  line = line:gsub('\r', '')
-  local parts = {}
-  local hls = {}
-  local pos = 1
-  local col = 0
-  local grp, grp_start
-  while pos <= #line do
-    local es, ee, params, code = line:find('\027%[([%d;]*)([A-Za-z])', pos)
-    if not es then
-      local chunk = line:sub(pos)
-      parts[#parts + 1] = chunk
-      col = col + #chunk
-      break
-    end
-    if es > pos then
-      local chunk = line:sub(pos, es - 1)
-      parts[#parts + 1] = chunk
-      col = col + #chunk
-    end
-    if code == 'm' then
-      if grp and col > grp_start then
-        hls[#hls + 1] = { col = grp_start, end_col = col, group = grp }
-      end
-      grp = nil
-      grp_start = nil
-      if params ~= '' and params ~= '0' then
-        for p in params:gmatch('(%d+)') do
-          local n = tonumber(p)
-          if n == 0 then
-            grp = nil
-            grp_start = nil
-          elseif n == 1 then
-            grp = grp or sgr_map[n]
-            grp_start = grp_start or col
-          elseif sgr_map[n] then
-            grp = sgr_map[n]
-            grp_start = col
-          end
-        end
-      end
-    end
-    pos = ee + 1
-  end
-  if grp and col > grp_start then
-    hls[#hls + 1] = { col = grp_start, end_col = col, group = grp }
-  end
-  return table.concat(parts), hls
-end
-
-local function offset_hls(hls, n)
-  for i, hl in ipairs(hls) do
-    hls[i] = { col = hl.col + n, end_col = hl.end_col + n, group = hl.group }
-  end
-  return hls
-end
-
-local function summary_status(prefix)
-  local labels = {
-    ['✓'] = 'success',
-    X = 'failure',
-    ['✗'] = 'failure',
-    ['-'] = 'skipped',
-    ['⊘'] = 'skipped',
-    ['~'] = 'in_progress',
-    ['●'] = 'in_progress',
-    ['*'] = 'queued',
-    OK = 'success',
-    FAIL = 'failure',
-    SKIP = 'skipped',
-    CANCEL = 'cancelled',
-    RUN = 'in_progress',
-    QUEUE = 'queued',
-  }
-  return labels[prefix]
-end
-
-local function summary_status_group(status)
-  local groups = {
-    success = 'ForgePass',
-    failure = 'ForgeFail',
-    skipped = 'ForgeLogDim',
-    cancelled = 'ForgeLogDim',
-    in_progress = 'ForgePending',
-    queued = 'ForgePending',
-  }
-  return groups[status] or 'ForgePending'
-end
-
-local function summary_status_icon(status)
-  local icons = config_mod.config().display.icons
-  status = (status or ''):lower()
-  if status == 'success' then
-    return icons.pass
-  end
-  if status == 'failure' or status == 'failed' then
-    return icons.fail
-  end
-  if
-    status == 'in_progress'
-    or status == 'running'
-    or status == 'pending'
-    or status == 'queued'
-  then
-    return icons.pending
-  end
-  if status == 'cancelled' or status == 'canceled' or status == 'skipped' then
-    return icons.skip
-  end
-  return icons.unknown
-end
-
-local function duration_label(duration)
-  if not duration or duration == '' then
-    return nil
-  end
-  return ('[%s]'):format(duration)
-end
-
----@param started string?
----@param completed string?
----@return string?
-local function format_duration(started, completed)
-  if not started or not completed then
-    return nil
-  end
-  local function parse_iso(s)
-    local y, mo, d, h, mi, sec = s:match('^(%d+)-(%d+)-(%d+)T(%d+):(%d+):(%d+)')
-    if not y then
-      return nil
-    end
-    return os.time({ year = y, month = mo, day = d, hour = h, min = mi, sec = sec })
-  end
-  local t0 = parse_iso(started)
-  local t1 = parse_iso(completed)
-  if not t0 or not t1 then
-    return nil
-  end
-  local diff = math.max(0, t1 - t0)
-  if diff < 60 then
-    return ('%ds'):format(diff)
-  elseif diff < 3600 then
-    return ('%dm %ds'):format(math.floor(diff / 60), diff % 60)
-  end
-  return ('%dh %dm'):format(math.floor(diff / 3600), math.floor(diff % 3600 / 60))
-end
-
----@class forge.StepEntry
----@field name string
----@field conclusion string
----@field started string?
----@field completed string?
----@field number integer
----@field emitted boolean
-
----@class forge.StepLookup
----@field by_name table<string, forge.StepEntry>
----@field ordered forge.StepEntry[]
----@field post_steps forge.StepEntry[]
----@field setup forge.StepEntry?
----@field complete forge.StepEntry?
-
----@param steps table[]?
----@return forge.StepLookup?
-local function build_step_lookup(steps)
-  if not steps or #steps == 0 then
-    return nil
-  end
-  table.sort(steps, function(a, b)
-    return a.number < b.number
-  end)
-  local lookup = {
-    by_name = {},
-    ordered = {},
-    post_steps = {},
-  }
-  for _, s in ipairs(steps) do
-    local entry = {
-      name = s.name,
-      conclusion = s.conclusion or '',
-      started = s.startedAt,
-      completed = s.completedAt,
-      number = s.number,
-      emitted = false,
-    }
-    lookup.ordered[#lookup.ordered + 1] = entry
-    lookup.by_name[s.name] = entry
-    if s.name == 'Set up job' then
-      lookup.setup = entry
-    elseif s.name == 'Complete job' then
-      lookup.complete = entry
-    elseif s.name:match('^Post ') then
-      lookup.post_steps[#lookup.post_steps + 1] = entry
-    end
-  end
-  return #lookup.ordered > 0 and lookup or nil
-end
-
----@param raw_lines string[]
----@param steps table[]?
----@return { lines: table[], headers: integer[], errors: integer[] }
-local function parse_github(raw_lines, steps)
-  local lines = {}
-  local headers = {}
-  local errors = {}
-  local cur_job, cur_step
-  local in_group = false
-  local lookup = build_step_lookup(steps)
-  local post_idx = 0
-  local unknown_step = false
-
-  ---@param se forge.StepEntry
-  local function emit_step(se)
-    if se.emitted then
-      return
-    end
-    se.emitted = true
-    in_group = false
-    local dur = format_duration(se.started, se.completed)
-    lines[#lines + 1] = {
-      text = se.name,
-      hls = {},
-      fold = '>2',
-      kind = 'step',
-      duration = dur,
-      conclusion = se.conclusion,
-    }
-    headers[#headers + 1] = #lines
-  end
-
-  for _, raw in ipairs(raw_lines) do
-    if raw:byte(1) == 0xEF and raw:byte(2) == 0xBB and raw:byte(3) == 0xBF then
-      raw = raw:sub(4)
-    end
-    raw = raw:gsub('\r', '')
-    local job, step, rest = raw:match('^([^\t]*)\t([^\t]*)\t(.*)$')
-    if not job then
-      local text, h = strip_ansi(raw)
-      lines[#lines + 1] = { text = text, hls = h, fold = '0', kind = 'raw' }
-      goto continue
-    end
-    if job ~= cur_job then
-      cur_job = job
-      cur_step = nil
-      post_idx = 0
-      in_group = false
-      lines[#lines + 1] = { text = job, hls = {}, fold = '>1', kind = 'job' }
-      headers[#headers + 1] = #lines
-      if lookup and lookup.setup then
-        emit_step(lookup.setup)
-      end
-    end
-    if not lookup and step ~= cur_step then
-      cur_step = step
-      in_group = false
-      unknown_step = step == 'UNKNOWN STEP'
-      if not unknown_step then
-        lines[#lines + 1] = { text = '  ' .. step, hls = {}, fold = '>2', kind = 'step' }
-        headers[#headers + 1] = #lines
-      end
-    end
-    do
-      local _, content = rest:match('^(%d%d%d%d%-%d%d%-%d%dT[%d:.]+Z)%s(.*)$')
-      if not content then
-        content = rest
-      end
-
-      if lookup then
-        if content == 'Post job cleanup.' then
-          post_idx = post_idx + 1
-          if lookup.post_steps[post_idx] then
-            emit_step(lookup.post_steps[post_idx])
-          end
-        elseif content:match('^Cleaning up orphan processes') then
-          if lookup.complete then
-            emit_step(lookup.complete)
-          end
-        end
-      end
-
-      local kind, m = 'content', nil
-      m = content:match('^##%[error[^%]]*%](.*)$')
-      if m then
-        kind = 'error'
-        content = 'Error: ' .. m
-        goto emit
-      end
-      m = content:match('^##%[warning[^%]]*%](.*)$')
-      if m then
-        kind = 'warning'
-        content = 'Warning: ' .. m
-        goto emit
-      end
-      m = content:match('^##%[group%](.*)$')
-      if m then
-        if lookup and lookup.by_name[m] then
-          emit_step(lookup.by_name[m])
-        end
-        kind = 'group'
-        content = m
-        goto emit
-      end
-      if content:match('^##%[endgroup') then
-        in_group = false
-        goto continue
-      end
-      m = content:match('^##%[debug[^%]]*%](.*)$')
-      if m then
-        kind = 'debug'
-        content = m
-        goto emit
-      end
-      m = content:match('^##%[notice[^%]]*%](.*)$')
-      if m then
-        kind = 'notice'
-        content = m
-        goto emit
-      end
-      m = content:match('^%[command%](.*)$')
-      if m then
-        kind = 'command'
-        content = m
-      end
-      ::emit::
-      local text, h = strip_ansi(content)
-      if kind == 'error' or kind == 'warning' then
-        h = {}
-      end
-      local indent_n
-      local fold_val
-      if lookup then
-        if kind == 'group' then
-          indent_n = 2
-          fold_val = '>3'
-        elseif in_group then
-          indent_n = 4
-          fold_val = '3'
-        else
-          indent_n = 2
-          fold_val = '2'
-        end
-      else
-        indent_n = (unknown_step and not in_group) and 2 or 4
-        if kind == 'group' then
-          fold_val = unknown_step and '>2' or '>3'
-        elseif unknown_step then
-          fold_val = in_group and '2' or '1'
-        elseif in_group then
-          fold_val = '3'
-        else
-          fold_val = '2'
-        end
-      end
-      lines[#lines + 1] = {
-        text = (' '):rep(indent_n) .. text,
-        hls = offset_hls(h, indent_n),
-        fold = fold_val,
-        kind = kind,
-      }
-      if kind == 'error' then
-        errors[#errors + 1] = #lines
-      elseif kind == 'group' then
-        headers[#headers + 1] = #lines
-        in_group = true
-      end
-    end
-    ::continue::
-  end
-  if lookup then
-    for _, se in ipairs(lookup.ordered) do
-      emit_step(se)
-    end
-  end
-  return { lines = lines, headers = headers, errors = errors }
-end
-
-local function parse_gitlab(raw_lines)
-  local lines = {}
-  local headers = {}
-  local errors = {}
-  local in_section = false
-
-  local function process_section_start(raw)
-    local _, sep_end = raw:find('\027%[0K')
-    local header_raw
-    if sep_end then
-      header_raw = raw:sub(sep_end + 1)
-    else
-      header_raw = raw:match('^section_start:%d+:(.+)$') or ''
-    end
-    local text, h = strip_ansi(header_raw)
-    if text == '' then
-      text = raw:match('^section_start:%d+:([^%c\027]+)') or 'section'
-    end
-    lines[#lines + 1] = { text = text, hls = h, fold = '>1', kind = 'section' }
-    headers[#headers + 1] = #lines
-    in_section = true
-  end
-
-  local function process_content(raw)
-    local text, h = strip_ansi(raw)
-    if text == '' then
-      return
-    end
-    local prefix = in_section and '  ' or ''
-    if #prefix > 0 then
-      offset_hls(h, #prefix)
-    end
-    local kind = 'content'
-    for _, hl in ipairs(h) do
-      if hl.group == 'ForgeFail' then
-        kind = 'error'
-        break
-      end
-    end
-    lines[#lines + 1] = {
-      text = prefix .. text,
-      hls = h,
-      fold = in_section and '1' or '0',
-      kind = kind,
-    }
-    if kind == 'error' then
-      errors[#errors + 1] = #lines
-    end
-  end
-
-  for _, raw in ipairs(raw_lines) do
-    raw = raw:gsub('\r', '')
-    if raw:match('^section_start:') then
-      process_section_start(raw)
-    elseif raw:match('^section_end:') then
-      in_section = false
-      local _, sep_end = raw:find('\027%[0K')
-      local remainder = sep_end and raw:sub(sep_end + 1) or ''
-      if remainder:match('^section_start:') then
-        process_section_start(remainder)
-      elseif remainder ~= '' then
-        process_content(remainder)
-      end
-    else
-      process_content(raw)
-    end
-  end
-  return { lines = lines, headers = headers, errors = errors }
-end
-
-local parser_for = {
-  github = parse_github,
-  codeberg = parse_github,
-  gitlab = parse_gitlab,
-}
-
-local function fold_ranges(parsed)
-  local ranges = {}
-  local stack = {}
-  for i, line in ipairs(parsed.lines) do
-    local expr = line.fold
-    if expr:sub(1, 1) == '>' then
-      local level = tonumber(expr:sub(2))
-      while #stack > 0 and stack[#stack].level >= level do
-        local f = table.remove(stack)
-        if i - 1 > f.start then
-          ranges[#ranges + 1] = { f.start, i - 1, f.level }
-        end
-      end
-      stack[#stack + 1] = { start = i, level = level }
-    else
-      local level = tonumber(expr) or 0
-      while #stack > 0 and stack[#stack].level > level do
-        local f = table.remove(stack)
-        if i - 1 > f.start then
-          ranges[#ranges + 1] = { f.start, i - 1, f.level }
-        end
-      end
-    end
-  end
-  local n = #parsed.lines
-  while #stack > 0 do
-    local f = table.remove(stack)
-    if n > f.start then
-      ranges[#ranges + 1] = { f.start, n, f.level }
-    end
-  end
-  table.sort(ranges, function(a, b)
-    return a[3] > b[3]
-  end)
-  return ranges
-end
-
-local function render(buf, parsed)
-  local texts = {}
-  for i, line in ipairs(parsed.lines) do
-    texts[i] = line.text
-  end
-  vim.bo[buf].modifiable = true
-  vim.api.nvim_buf_set_lines(buf, 0, -1, false, texts)
-  vim.bo[buf].modifiable = false
-  vim.api.nvim_buf_clear_namespace(buf, ns, 0, -1)
-  for i, line in ipairs(parsed.lines) do
-    local lnum = i - 1
-    if line.kind == 'job' then
-      vim.api.nvim_buf_set_extmark(buf, ns, lnum, 0, { line_hl_group = 'ForgeLogJob' })
-    elseif line.kind == 'step' then
-      local hl = line.conclusion == 'failure' and 'ForgeFail' or 'ForgeLogStep'
-      local ext_opts = { line_hl_group = hl }
-      if line.duration then
-        ext_opts.virt_text = { { duration_label(line.duration), 'ForgeLogDim' } }
-        ext_opts.virt_text_pos = 'eol'
-      end
-      vim.api.nvim_buf_set_extmark(buf, ns, lnum, 0, ext_opts)
-    elseif line.kind == 'group' then
-      vim.api.nvim_buf_set_extmark(buf, ns, lnum, 0, { line_hl_group = 'ForgeLogStep' })
-    elseif line.kind == 'section' then
-      vim.api.nvim_buf_set_extmark(buf, ns, lnum, 0, { line_hl_group = 'ForgeLogSection' })
-    elseif line.kind == 'error' then
-      vim.api.nvim_buf_set_extmark(buf, ns, lnum, 0, { line_hl_group = 'ForgeLogError' })
-      local start = line.text:find('Error: ', 1, true)
-      if start then
-        vim.api.nvim_buf_set_extmark(buf, ns, lnum, start - 1, {
-          end_col = start + 6,
-          hl_group = 'ForgeLogErrorLabel',
-        })
-      end
-    elseif line.kind == 'warning' then
-      vim.api.nvim_buf_set_extmark(buf, ns, lnum, 0, { line_hl_group = 'ForgeLogWarning' })
-      local start = line.text:find('Warning: ', 1, true)
-      if start then
-        vim.api.nvim_buf_set_extmark(buf, ns, lnum, start - 1, {
-          end_col = start + 8,
-          hl_group = 'ForgeLogWarningLabel',
-        })
-      end
-    elseif line.kind == 'notice' then
-      vim.api.nvim_buf_set_extmark(buf, ns, lnum, 0, { line_hl_group = 'ForgeLogWarning' })
-    elseif line.kind == 'command' then
-      vim.api.nvim_buf_set_extmark(buf, ns, lnum, 0, {
-        end_col = #line.text,
-        hl_group = 'ForgeLogCommand',
-      })
-    elseif line.kind == 'debug' then
-      vim.api.nvim_buf_set_extmark(buf, ns, lnum, 0, { line_hl_group = 'ForgeLogDim' })
-    end
-    for _, hl in ipairs(line.hls) do
-      vim.api.nvim_buf_set_extmark(buf, ns, lnum, hl.col, {
-        end_col = hl.end_col,
-        hl_group = hl.group,
-      })
-    end
-  end
-  local fold_meta = {}
-  for i, line in ipairs(parsed.lines) do
-    if line.kind == 'step' and line.duration then
-      fold_meta[i] = { duration = line.duration, conclusion = line.conclusion }
-    end
-  end
-  set_data(buf, {
-    headers = parsed.headers,
-    errors = parsed.errors,
-    fold_meta = fold_meta,
-  })
-  local ranges = fold_ranges(parsed)
-  local wins = vim.fn.win_findbuf(buf)
-  if #wins > 0 then
-    vim.api.nvim_win_call(wins[1], function()
-      vim.wo[0].foldmethod = 'manual'
-      vim.wo[0].foldtext = 'v:lua.require("forge.log")._foldtext()'
-      pcall(vim.cmd, 'silent! normal! zE')
-      for _, r in ipairs(ranges) do
-        vim.cmd(r[1] .. ',' .. r[2] .. 'fold')
-      end
-      vim.wo[0].foldlevel = 99
-    end)
-  end
-end
+local strip_ansi = log_render.strip_ansi
+local duration_label = log_render.duration_label
+local parse_github = log_render.parse_github
+local parse_gitlab = log_render.parse_gitlab
+local parser_for = log_render.parser_for
+local fold_meta = log_render.fold_meta
 
 local function jump(buf, kind, dir)
   local d = buf_data[buf]
@@ -938,7 +354,12 @@ function M.open(cmd, opts, reuse_buf)
         raw_lines[#raw_lines] = nil
       end
       local parsed = parse(raw_lines, steps)
-      render(buf, parsed)
+      set_data(buf, {
+        headers = parsed.headers,
+        errors = parsed.errors,
+        fold_meta = fold_meta(parsed),
+      })
+      log_render.render(buf, ns, parsed)
       local wins = vim.fn.win_findbuf(buf)
       if #wins > 0 then
         local win = wins[1]
@@ -1006,210 +427,11 @@ end
 ---@field browse_url_fn fun(job_id: string): string?
 ---@field log_cmd_fn fun(job_id: string, failed: boolean): string[], forge.LogOpts
 
-local function normalize_job_filter_input(text)
-  if text == nil then
-    return nil, true
-  end
-  local trimmed = vim.trim(text)
-  if trimmed == '' then
-    return nil, false
-  end
-  return trimmed, false
-end
-
-local function matches_job_filter(name, job_filter)
-  if type(job_filter) ~= 'string' or job_filter == '' then
-    return true
-  end
-  local candidate = type(name) == 'string' and strip_ansi(name):lower() or ''
-  return candidate:find(job_filter:lower(), 1, true) ~= nil
-end
-
-local function filter_summary(parsed, job_filter)
-  if type(job_filter) ~= 'string' or job_filter == '' then
-    return parsed
-  end
-  local matched_ids = {}
-  local header_rows = {}
-  for _, lnum in ipairs(parsed.job_lnums or {}) do
-    header_rows[lnum] = true
-    local job = parsed.jobs and parsed.jobs[lnum] or nil
-    if job and job.id and matches_job_filter(parsed.lines[lnum], job_filter) then
-      matched_ids[job.id] = true
-    end
-  end
-  local lines = {}
-  local hls = {}
-  local jobs = {}
-  local job_lnums = {}
-  local first_job = parsed.job_lnums and parsed.job_lnums[1] or nil
-  local header_limit = first_job and (first_job - 1) or 0
-  for lnum = 1, header_limit do
-    lines[#lines + 1] = parsed.lines[lnum]
-    hls[#lines] = parsed.hls[lnum] or {}
-  end
-  if next(matched_ids) == nil then
-    local text = ('No jobs matching "%s"'):format(job_filter)
-    lines[#lines + 1] = text
-    hls[#lines] = {
-      {
-        col = 0,
-        end_col = #text,
-        group = 'ForgeDim',
-      },
-    }
-    return { lines = lines, hls = hls, jobs = jobs, job_lnums = job_lnums }
-  end
-  for lnum, line in ipairs(parsed.lines or {}) do
-    local job = parsed.jobs and parsed.jobs[lnum] or nil
-    if job and matched_ids[job.id] then
-      lines[#lines + 1] = line
-      hls[#lines] = parsed.hls[lnum] or {}
-      jobs[#lines] = job
-      if header_rows[lnum] then
-        job_lnums[#job_lnums + 1] = #lines
-      end
-    end
-  end
-  return { lines = lines, hls = hls, jobs = jobs, job_lnums = job_lnums }
-end
-
-local function parse_summary(raw_lines)
-  local lines = {}
-  local hls = {}
-  local jobs = {}
-  local job_lnums = {}
-  local section
-  local current_job
-  local first = 1
-  local last = #raw_lines
-
-  while first <= last do
-    local text = strip_ansi(raw_lines[first])
-    if not text:match('^%s*$') then
-      break
-    end
-    first = first + 1
-  end
-
-  while last >= first do
-    local text = strip_ansi(raw_lines[last])
-    if not text:match('^%s*$') then
-      break
-    end
-    last = last - 1
-  end
-
-  for i = first, last do
-    local raw = raw_lines[i]
-    local text, h = strip_ansi(raw)
-    local lnum = #lines + 1
-    if text:match('^%s*$') then
-      text = ''
-      h = {}
-      current_job = nil
-    else
-      local prefix = text:match('^(%S+)')
-      local status = prefix and summary_status(prefix) or nil
-      if text:match('^%u[%u%s]+$') then
-        section = text
-        current_job = nil
-      end
-      local job_id = text:match('%(ID (%d+)%)%s*$')
-      if job_id then
-        jobs[lnum] = { id = job_id, failed = status == 'failure' }
-        current_job = jobs[lnum]
-        job_lnums[#job_lnums + 1] = lnum
-      elseif current_job and text:match('^%s+') and section ~= 'ANNOTATIONS' then
-        jobs[lnum] = current_job
-      end
-    end
-    lines[lnum] = text
-    hls[lnum] = h
-  end
-
-  return { lines = lines, hls = hls, jobs = jobs, job_lnums = job_lnums }
-end
-
-local function summary_job_at_line(raw_lines, lnum)
-  local first, last = 1, #raw_lines
-
-  while first <= last do
-    local text = strip_ansi(raw_lines[first])
-    if not text:match('^%s*$') then
-      break
-    end
-    first = first + 1
-  end
-
-  while last >= first do
-    local text = strip_ansi(raw_lines[last])
-    if not text:match('^%s*$') then
-      break
-    end
-    last = last - 1
-  end
-
-  if lnum < first or lnum > last then
-    return nil
-  end
-
-  return parse_summary(raw_lines).jobs[lnum - first + 1]
-end
-
-local function summary_hls(prefix, status)
-  return { { col = 0, end_col = #prefix, group = summary_status_group(status) } }
-end
-
-local function parse_summary_json(data)
-  local lines = {}
-  local hls_list = {}
-  local jobs = {}
-  local job_lnums = {}
-
-  local run_status = (data.conclusion and data.conclusion ~= '') and data.conclusion or data.status
-  local header_prefix = summary_status_icon(run_status) .. '  '
-  local header = header_prefix .. (data.displayTitle or data.name or 'run')
-  lines[#lines + 1] = header
-  hls_list[#hls_list + 1] = summary_hls(header_prefix, run_status)
-
-  lines[#lines + 1] = ''
-  hls_list[#hls_list + 1] = {}
-
-  for _, job in ipairs(data.jobs or {}) do
-    local jstatus = (job.conclusion and job.conclusion ~= '') and job.conclusion or job.status
-    local prefix = summary_status_icon(jstatus) .. '  '
-    local line = prefix .. (job.name or 'job')
-    local dur = duration_label(format_duration(job.startedAt, job.completedAt))
-    if dur then
-      line = line .. '  ' .. dur
-    end
-    if jstatus == 'in_progress' and job.steps then
-      local done, total = 0, 0
-      for _, s in ipairs(job.steps) do
-        total = total + 1
-        local sc = (s.conclusion and s.conclusion ~= '') and s.conclusion or s.status
-        if sc == 'success' or sc == 'failure' or sc == 'skipped' then
-          done = done + 1
-        end
-      end
-      if total > 0 then
-        line = line .. ('  [%d/%d]'):format(done, total)
-      end
-    end
-    lines[#lines + 1] = line
-    hls_list[#hls_list + 1] = summary_hls(prefix, jstatus)
-
-    local job_id = tostring(job.databaseId or '')
-    if job_id ~= '' then
-      local failed = jstatus == 'failure'
-      jobs[#lines] = { id = job_id, failed = failed }
-      job_lnums[#job_lnums + 1] = #lines
-    end
-  end
-
-  return { lines = lines, hls = hls_list, jobs = jobs, job_lnums = job_lnums }
-end
+local normalize_job_filter_input = log_summary.normalize_job_filter_input
+local filter_summary = log_summary.filter_summary
+local parse_summary = log_summary.parse_summary
+local summary_job_at_line = log_summary.summary_job_at_line
+local parse_summary_json = log_summary.parse_summary_json
 
 ---@param cmd string[]
 ---@param opts forge.SummaryOpts
